refactor(internal): merge internal package
