+++ cmd/check/check_command.go
@@ -3,17 +3,17 @@ package main
 import (
 	"time"
 
-	"github.com/henry40408/ssh-shell-resource/internal/models"
+	"github.com/henry40408/ssh-shell-resource/internal"
 )
 
-func CheckCommand(request *models.CheckRequest) models.CheckResponse {
-	versions := models.CheckResponse{}
+func CheckCommand(request *internal.CheckRequest) internal.CheckResponse {
+	versions := internal.CheckResponse{}
 
 	previousVersion := request.Version
 	if !previousVersion.Timestamp.IsZero() {
 		versions = append(versions, previousVersion)
 	}
 
-	versions = append(versions, models.Version{Timestamp: time.Now()})
+	versions = append(versions, internal.Version{Timestamp: time.Now()})
 	return versions
 }
